Extract applySubset functionality from plywood route
